Move helpers from deprecated trait to HasFactory trait
